chore: Split out streams (#3)
* move streams to subdir

* remove dynamic streams module

* add init to allow imports to work the same

* add missing stream
