+++ tap_googleads/dynamic_streams/click_view_report.py
@@ -1,0 +1,112 @@
+"""ClickViewReportStream for Google Ads tap."""
+
+from __future__ import annotations
+
+import datetime
+from http import HTTPStatus
+from typing import TYPE_CHECKING, Any, Dict
+
+from tap_googleads.client import ResumableAPIError
+from tap_googleads.dynamic_query_stream import DynamicQueryStream
+
+if TYPE_CHECKING:
+    from singer_sdk.helpers.types import Context
+
+
+class ClickViewReportStream(DynamicQueryStream):
+    """Stream for click view reports."""
+    
+    date: datetime.date
+
+    @property
+    def gaql(self):
+        return f"""
+        SELECT
+            click_view.gclid
+            , customer.id
+            , click_view.ad_group_ad
+            , ad_group.id
+            , ad_group.name
+            , campaign.id
+            , campaign.name
+            , segments.ad_network_type
+            , segments.device
+            , segments.date
+            , segments.slot
+            , metrics.clicks
+            , segments.click_type
+            , click_view.keyword
+            , click_view.keyword_info.match_type
+        FROM click_view
+        WHERE segments.date = '{self.date.isoformat()}'
+        """
+
+    name = "click_view_report"
+    primary_keys = [
+        "clickView__gclid",
+        "clickView__keyword",
+        "clickView__keywordInfo__matchType",
+        "customer__id",
+        "adGroup__id",
+        "campaign__id",
+        "segments__device",
+        "segments__adNetworkType",
+        "segments__slot",
+        "date",
+    ]
+    replication_key = "date"
+
+    def post_process(self, row, context):
+        row["date"] = row["segments"].pop("date")
+
+        if row.get("clickView", {}).get("keyword") is None:
+            row["clickView"]["keyword"] = "null"
+            row["clickView"]["keywordInfo"] = {"matchType": "null"}
+
+        return super().post_process(row, context)
+
+    def get_url_params(self, context, next_page_token):
+        """Return a dictionary of values to be used in URL parameterization.
+
+        Args:
+            context: The stream context.
+            next_page_token: The next page index or value.
+
+        Returns:
+            A dictionary of URL query parameters.
+
+        """
+        params: dict = {}
+        if next_page_token:
+            params["pageToken"] = next_page_token
+        return params
+
+    def request_records(self, context):
+        start_value = self.get_starting_replication_key_value(context)
+
+        start_date = datetime.date.fromisoformat(start_value)
+        end_date = datetime.date.fromisoformat(self.config["end_date"])
+
+        delta = end_date - start_date
+        dates = (start_date + datetime.timedelta(days=i) for i in range(delta.days))
+
+        for self.date in dates:
+            records = list(super().request_records(context))
+
+            if not records:
+                self._increment_stream_state(
+                    {"date": self.date.isoformat()}, context=self.context
+                )
+
+            yield from records
+
+    def validate_response(self, response):
+        if response.status_code == HTTPStatus.FORBIDDEN:
+            error = response.json()["error"]["details"][0]["errors"][0]
+            msg = (
+                "Click view report not accessible to customer "
+                f"'{self.context['customer_id']}': {error['message']}"
+            )
+            raise ResumableAPIError(msg, response)
+
+        super().validate_response(response)
